% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching.R
\name{abbreviated}
\alias{abbreviated}
\title{Checks if two country names are abbreviations of each other.}
\usage{
abbreviated(str1, str2)
}
\arguments{
\item{str1}{A string representing the first country name.}

\item{str2}{A string representing the second country name.}
}
\value{
TRUE if the two country names are abbreviations of each other;
otherwise, FALSE.
}
\description{
Checks if two country names are abbreviations of each other.
}
\examples{
is_abbreviation <- abbreviated("us", "united states")
}
