% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bar_charts.R
\name{add_bar_charts}
\alias{add_bar_charts}
\title{Add a Bar Chart Layer to a ggplot layer}
\usage{
add_bar_charts(
  df,
  width,
  height,
  attributes,
  label_x = "label_x",
  label_y = "label_y"
)
}
\arguments{
\item{df}{A data frame containing the data to be plotted.}

\item{width}{The width of the bar charts}

\item{height}{The height of the bar charts}

\item{attributes}{A list of column names in \code{df} that will be used to
generate the bars.}

\item{label_x}{The name of the column in \code{df} that contains the
x-coordinate for positioning the bar charts. Default is \code{"label_x"}.}

\item{label_y}{The name of the column in \code{df} that contains the
y-coordinate for positioning the bar charts. Default is \code{"label_y"}.}
}
\value{
A layer of barcharts which can be added to a ggplot object
}
\description{
Add a Bar Chart Layer to a ggplot layer
}
\examples{
df <- data.frame(id = 1:5, value1 = c(5, 4, 3, NA, 2),
 value2 = c(2, 3, 4, 5, 6), label_x = 1:5, label_y = rep(1, 5))
map <- map + add_bar_charts(df, width = 100,
 height = 50, attributes = c("value1", "value2"))

}
