% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lines.R
\name{add_lines_to_labels}
\alias{add_lines_to_labels}
\title{Adds lines connecting small countries to their closest rectangles}
\usage{
add_lines_to_labels(data)
}
\arguments{
\item{data}{The map data which should include simple features information.}

\item{width}{The width of the free space the small countries
should be assigned}

\item{height}{The height of the free space the small countries
should be assigned}
}
\value{
A geom_segment which contains the lines between countries and
the closest rectangles
}
\description{
Adds lines connecting small countries to their closest rectangles
}
\examples{
# Assuming `world_data` is a valid `sf` object containing world map data:
world_data <- modify_label_positions(world_data,
                                     small_country_area = 100000,
                                     width = 1, height = 1) +
plot_with_lines <- ggplot() +
  geom_sf(data = world_data) +
  add_lines_to_labels(data = world_data)

}
