% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pie_charts.R
\name{add_pie_charts}
\alias{add_pie_charts}
\title{Add Pie Charts to a Plot}
\usage{
add_pie_charts(data, x, y, attributes, legend_title = NULL, pie_scale = 1)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{x}{A  string indicating the name of the column in \code{data} that contains
the x-coordinate for positioning the pie charts.}

\item{y}{A string indicating the name of the column in \code{data} that contains
the y-coordinate for positioning the pie charts.}

\item{attributes}{A list of column names in \code{data} that will be used to
fill the pie charts.}

\item{legend_title}{A string to be used as the title of the legend.}

\item{pie_scale}{The size of the pie charts}
}
\value{
A \code{ggplot} object containing the pie layer
}
\description{
Generates a pie chart layer for a ggplot object
}
\examples{
data <- data.frame(x = c(1, 2, 3), y = c(4, 5, 6),
 A = c(30, 50, 20), B = c(70, 50, 80))
map <- map + add_pie_charts(data, x = "x", y = "y",
 attributes = c("A", "B"), legend_title = "Category", pie_scale = 0.1)

}
