% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_points.R
\name{add_points}
\alias{add_points}
\title{Generates a point layer based off an attribute}
\usage{
add_points(data, x, y, attribute, legend_title)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{x}{The name of the column in \code{data} that contains the x-coordinates for
the points.}

\item{y}{The name of the column in \code{data} that contains the y-coordinates for
the points.}

\item{attribute}{The name of the column in \code{data} that contains the variable
to determine the size and color of the points.}

\item{legend_title}{The title of the legend.}
}
\value{
A layer of points that can be added to a ggplot object.
}
\description{
Generates a point layer based off an attribute
}
\examples{
df <- data.frame(x = rnorm(100), y = rnorm(100), measure = rnorm(100))
plot <- ggplot() + add_points(df, x = "x", y = "y",
 attribute = "attribute", legend_title = "Attribute")

}
