% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bar_charts.R
\name{build_layer}
\alias{build_layer}
\title{Builds a bar chart}
\usage{
build_layer(df, width, height, attributes, label_x, label_y)
}
\arguments{
\item{df}{A data frame containing the coordinates for positioning the
bar chart.}

\item{width}{The width of the bar chart.}

\item{height}{The height of the bar chart.}

\item{label_x}{The name of the column in \code{df} that contains the x-coordinate
for positioning the bar chart.}

\item{label_y}{The name of the column in \code{df} that contains the y-coordinate
for positioning the bar chart.}

\item{data}{A data frame with the bar chart information.}
}
\value{
The bar chart as a geom object
}
\description{
Builds a bar chart
}
\examples{
# Sample data frame
df <- data.frame(label_x = 5, label_y = 5)
attributes <- c("value1", "value2")
df_data <- data.frame(value1 = 10, value2 = 20)
bar_layer <- build_layer(df = df_data,
                         width = 100,
                         height = 50,
                         attributes = attributes,
                         label_x = "label_x",
                         label_y = "label_y")


}
