% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stars.R
\name{build_star_layer}
\alias{build_star_layer}
\title{Builds an individual star layer}
\usage{
build_star_layer(df, data, width, height, star_size, label_x, label_y)
}
\arguments{
\item{df}{A data frame containing the coordinates for positioning the
star rating layer.}

\item{data}{A data frame with columns \code{x}, \code{y}, and \code{colours} which represent
the positions and color of the stars (yellow for filled, grey for empty).}

\item{width}{The width of the star layer}

\item{height}{The height of the star layer}

\item{star_size}{The size of each individual star}

\item{label_x}{The name of the column in \code{df} that contains the x-coordinate
for positioning the star ratings.}

\item{label_y}{The name of the column in \code{df} that contains the y-coordinate
for positioning the star ratings.}
}
\value{
The star layer as a geom object
}
\description{
Builds an individual star layer
}
\examples{
df <- data.frame(label_x = 5, label_y = 5)
data <- data.frame(x = 1:5, y = rep(1, 5),
 colours = c("yellow", "yellow", "yellow", "grey", "grey"))
build_star_layer(df = df, data = data,
 width = 100, height = 50, star_size = 10, label_x = "label_x",
 label_y = "label_y")

}
