% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching.R
\name{compare_nearest_country}
\alias{compare_nearest_country}
\title{Compares if two countries are the same based on their Levenshtein distance.}
\usage{
compare_nearest_country(country1, country2, max_distance)
}
\arguments{
\item{country1}{A string representing the first country name.}

\item{country2}{A string representing the second country name.}

\item{max_distance}{A numeric threshold for the maximum allowed
Levenshtein distance.}
}
\value{
TRUE if both countries are the same based on the
Levenshtein distance; otherwise, FALSE.
}
\description{
Compares if two countries are the same based on their Levenshtein distance.
}
\examples{
are_same <- compare_nearest_country("germany", "germnay", 3)
}
