% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{convert_columns_to_number}
\alias{convert_columns_to_number}
\title{Converts specified columns to numeric values after removing
unwanted characters.}
\usage{
convert_columns_to_number(data, columns, characters_to_remove)
}
\arguments{
\item{data}{A data frame containing the columns to convert.}

\item{columns}{A vector of column names to be converted to a number.}

\item{characters_to_remove}{A vector of characters to remove from the
columns before conversion to a number.}
}
\value{
A data frame with the specified columns converted to numeric values.
}
\description{
Converts specified columns to numeric values after removing
unwanted characters.
}
\examples{
converted_data <- convert_columns_to_number(data,
 columns = c("population", "gdp"), characters_to_remove = c("$", "\%"))

}
