% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching.R
\name{create_country_map}
\alias{create_country_map}
\title{Creates a mapping from country names to numbers}
\usage{
create_country_map()
}
\value{
A map of country names to numbers
}
\description{
Creates a mapping from country names to numbers
}
\examples{
country_map <- create_country_map()
}
