% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{create_map}
\usage{
create_map(
  sf = NULL,
  sfby = NULL,
  sfby.code = NULL,
  bbox = NULL,
  stats = NULL,
  by = NULL,
  variable,
  digits = 2,
  lang = "en",
  maptype = "choropleth",
  classtype = "jenks",
  classnumber = 5,
  classints = NULL,
  breaks,
  col = "black",
  pal = NULL,
  invertpal = FALSE,
  bgCol = "transparent",
  bgBorderCol = "transparent",
  waterbCol = bgCol,
  displayWaterboundaries = TRUE,
  faoareas = FALSE,
  faoareasLwd = 1,
  faoareasCol = "blue",
  faoareasLabels = TRUE,
  un_sdg_maptype = "nocolor",
  boundCol = "white",
  contCol = "lightgray",
  hashCol = "lightgray",
  m49_codes_to_hide = "010",
  add_small_features_as_dots = TRUE,
  add_small_NA_features_as_dots = FALSE,
  small_features_dots_cex = 0.4,
  pch = 21,
  level.min = NULL,
  level.max = NULL,
  level.factor = 1,
  level.unit = "chars",
  plot.type = "plot",
  plot.locator = "point_on_surface",
  plot.handler = NULL,
  plot.size = c(0.15, 0.15),
  legend = TRUE,
  legendtitle = "Legend",
  legendunit = "",
  legendcol = "black",
  legendpch = pch,
  legendcex = 0.8,
  legendpchcol = "black",
  legend_items = NULL,
  legend_nesting = FALSE,
  halo = FALSE,
  halocol = "black",
  halolwd = 1,
  naCol = "gray",
  naColBox = "lightgray",
  naLabel = "No Data",
  naHashCol = NULL,
  naHashLwd = 0.1,
  naHashDensity = 50,
  add_disclaimers = TRUE,
  add_copyright = TRUE,
  add = FALSE,
  debug = FALSE,
  family = "FuturaStd"
)
}
\description{
Creates a map
}
