% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map_background.R
\name{create_map_background}
\alias{create_map_background}
\title{create_map_background}
\usage{
create_map_background(
  bbox = NULL,
  bgCol = "transparent",
  bgBorderCol = "lightgray"
)
}
\description{
Creates a map background
}
