% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_with_classes.R
\name{enrich_with_classes}
\alias{enrich_with_classes}
\title{enrich_with_classes}
\usage{
enrich_with_classes(sf, classints, variable, maptype, level.min, level.max, level.unit)
}
\value{
an object from \pkg{sf}
}
\description{
Enrichs a spatialized statistical dataset with classes
}
