% Generated by roxygen2 (4.0.0): do not edit by hand
\name{geo_code}
\alias{geo_code}
\title{Geocodign API Wrapper}
\usage{
geo_code(street, city, state, zip, api.key)
}
\arguments{
\item{street}{A street portion of the address.}

\item{city}{A city.}

\item{state}{A state abbreviation.}

\item{zip}{A zip code.}

\item{api.key}{A \href{http://geoservices.tamu.edu/Services/Geocode/WebService/}{Texas A & M Geoservices API key}}
}
\value{
Returns a \code{data.frame} with latidue and longitude values.
}
\description{
A geocoding API that works with
\href{http://geoservices.tamu.edu/Services/Geocode/}{Texas A & M Geoservices}.
}
\note{
To receive an API key the user must
\href{http://geoservices.tamu.edu/Pricing/Partner/}{qualify} and then sing up: \url{https://geoservices.tamu.edu/Signup/}.
}
\author{
Bryan Goodrich and Tyler Rinker <tyler.rinker@gmail.com>.
}
\references{
\url{http://geoservices.tamu.edu/Services/Geocode/WebService/}
}
\keyword{geocode}

