% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baselayers.R
\name{get_baselayers}
\alias{get_baselayers}
\title{get_baselayers}
\usage{
get_baselayers()
}
\value{
an \code{list} of \pkg{sf} objects
}
\description{
Get UN base layers as \pkg{sf} objects based on UN-FAO Fisheries & Aquaculture Division
 Geoserver (using the OGC WFS protocol). Once retrieved, the layers will be cached internally to avoid
 re-fetching layers each time.
}
\examples{
\donttest{
  layers = get_baselayers()
}

}
