% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching.R
\name{get_country_indexes_from_dataframe}
\alias{get_country_indexes_from_dataframe}
\title{Returns the indexes of countries from a list of country names in a data frame}
\usage{
get_country_indexes_from_dataframe(
  countries,
  data_frame_countries,
  max_distance
)
}
\arguments{
\item{countries}{A vector of country names to search for.}

\item{data_frame_countries}{A vector of country names in the data frame.}

\item{max_distance}{A numeric threshold for the maximum allowed
Levenshtein distance.}
}
\value{
A list of indexes where the countries in \code{countries}
are found in the data frame.
}
\description{
Returns the indexes of countries from a list of country names in a data frame
}
\examples{
country_indexes <- get_country_indexes_from_dataframe(
 countries = c("Germany", "France"),
 data_frame_countries = csv_data$V1, max_distance = 3)
}
