% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching.R
\name{get_country_number}
\alias{get_country_number}
\title{Returns the country number based off the csv file}
\usage{
get_country_number(country, x)
}
\arguments{
\item{country}{A string representing the country name to search for.}

\item{x}{A numeric threshold for the Levenshtein distance.}
}
\value{
The row index of the closest matching country
if found; otherwise, -1.
}
\description{
This function first checks if a cached \code{country_map} exists and uses it for
fast lookup. If not, it will compute the mapping by calling
\code{create_country_map}.
The function tries to match the input \code{country} with names in the map.
It also attempts to clean the country name by removing common
words like "and", "of", or "the".
If a direct match isn't found, it uses the Levenshtein distance to search for
a close match.
}
\examples{
country_number <- get_country_number("germany", 3)
}
