% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching.R
\name{get_region_number}
\alias{get_region_number}
\title{Returns the country number based off the csv file}
\usage{
get_region_number(
  region,
  csv_path,
  distance = 3,
  method = "qgram",
  basic = FALSE
)
}
\arguments{
\item{region}{A string representing the region name to search for.}

\item{csv_path}{A string specifying the path of the csv which has
alternative names for regins in its rows}

\item{method}{A string representing the edit distance method}

\item{basic}{A boolean which sets whether edit distance is used or not}

\item{x}{A numeric threshold for the Levenshtein distance.}
}
\value{
The row index of the closest matching region
if found; otherwise, -1.
}
\description{
The function tries to match the input region with names in the csv.
It also attempts to clean the country name by removing common
words like "and", "of", or "the".
If a direct match isn't found, it uses the Levenshtein distance to search for
a close match.
}
\examples{
country_number <- get_region_number("germany", csv_path, 3)
}
