% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUNBaseLayers.R
\name{getUNBaseLayers}
\alias{getUNBaseLayers}
\title{getUNBaseLayers}
\usage{
getUNBaseLayers()
}
\value{
an \code{list} of \pkg{sf} objects
}
\description{
Get UN base layers as \pkg{sf} objects based on UN-FAO Fisheries & Aquaculture Division
 Geoserver (using the OGC WFS protocol).
}
\examples{
\donttest{
  layers = getUNBaseLayers()
}

}
