% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching.R
\name{is_acronym}
\alias{is_acronym}
\title{Checks if a name is an acronym of another.}
\usage{
is_acronym(str1, str2)
}
\arguments{
\item{str1}{A string representing the first name.}

\item{str2}{A string representing the second name.}
}
\value{
TRUE if the first string is an acronym of the second;
otherwise, FALSE.
}
\description{
Checks if a name is an acronym of another.
}
\examples{
acronym <- is_acronym("us", "united states")
}
