% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching.R
\name{is_shortened}
\alias{is_shortened}
\title{Checks if a name is a shortened version of another using a . to represent
a shortened string.}
\usage{
is_shortened(str1, str2)
}
\arguments{
\item{str1}{A string representing the first name.}

\item{str2}{A string representing the second name.}
}
\value{
TRUE if the first string is an shortened version of the second;
otherwise, FALSE.
}
\description{
Checks if a name is a shortened version of another using a . to represent
a shortened string.
}
\examples{
shortened <- is_shortened("un.", "united")
}
