% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching.R
\name{levenshtein_distance_lesser_than}
\alias{levenshtein_distance_lesser_than}
\title{Calculates the Levenshtein distance between two strings and checks
if it is less than a threshold.}
\usage{
levenshtein_distance_lesser_than(str1, str2, x)
}
\arguments{
\item{str1}{A string representing the first country name.}

\item{str2}{A string representing the second country name.}

\item{x}{A numeric threshold for the Levenshtein distance.}
}
\value{
TRUE if the Levenshtein distance is less than \code{x}; otherwise, FALSE.
}
\description{
Calculates the Levenshtein distance between two strings and checks
if it is less than a threshold.
}
\examples{
is_closer <- levenshtein_distance_lesser_than("germany", "germnay", 3)
}
