% Generated by roxygen2 (4.0.2): do not edit by hand
\name{map}
\alias{map}
\title{Easier choropleth maps}
\usage{
map(..., area = "Italy")
}
\arguments{
\item{...}{Arguments for the called function}

\item{area}{The geographical area for which the map should be drawn (currently only 'Italy')}
}
\description{
Easier choropleth maps
}
\author{
Nicola Sturaro
}

