\name{mapIT-package}
\alias{mapIT-package}

\docType{package}
\title{
Choropleth Maps of Italy
}
\description{
Creazione di mappe (coroplete) delle regioni italiane in maniera semplice.
}
\details{
\tabular{ll}{
Package: \tab mapIT\cr
Type: \tab Package\cr
Version: \tab 0.1.3\cr
Date: \tab 2014-03-21\cr
License: \tab LGPL-3\cr
}
}
\author{
Nicola Sturaro Sommacal

Maintainer: Nicola Sturaro Sommacal <nicola.sturaro@quantide.com>

}


\keyword{ package }

\examples{

# esempio ultra-minimale. per la guida ed esempi si veda \code{\link{mapIT}}
mapIT(rnorm(20)) 


}
