% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Merges two datasets based on region names.}
\usage{
merge_data(region_data, auxiliary_data, region_name, auxiliary_region_name)
}
\arguments{
\item{region_data}{A data frame containing the region data.}

\item{auxiliary_data}{A data frame containing the auxiliary data to merge.}

\item{region_name}{A string specifying the column name in \code{region}
that contains the region names.}

\item{auxiliary_region_name}{A string specifying the column name in
\code{auxiliary_data} that contains the region names}
}
\value{
An \code{sf} data frame containing the merged data.
}
\description{
Merges two datasets based on region names.
}
\examples{
merged_data <- merge_data(country_data, auxiliary_data,
                          "country", "country_name")
}
