% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data_with_country_matching}
\alias{merge_data_with_country_matching}
\title{Merges two datasets based on country names, using name matching.}
\usage{
merge_data_with_country_matching(
  country_data,
  auxillary_data,
  country_name,
  auxillary_country_name
)
}
\arguments{
\item{country_data}{A data frame containing the country data.}

\item{auxillary_data}{A data frame containing the auxiliary data to merge.}

\item{country_name}{A string specifying the column name in \code{country_data}
that contains the country names.}

\item{auxillary_country_name}{A string specifying the column name in
\code{auxillary_data} that contains the country names}
}
\value{
An \code{sf} data frame containing the merged data.
}
\description{
Merges two datasets based on country names, using name matching.
}
\examples{
merged_data <- merge_data_with_country_matching(country_data, auxillary_data,
                                                "country", "country_name")
}
