% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data_with_csv}
\alias{merge_data_with_csv}
\title{Merges two datasets based on values in a csv.}
\usage{
merge_data_with_csv(
  region_data,
  auxiliary_data,
  region_name,
  auxiliary_region_name,
  csv_path,
  distance = 1,
  method = "qgram",
  basic = FALSE
)
}
\arguments{
\item{region_data}{A data frame containing the region data.}

\item{auxiliary_data}{A data frame containing the auxiliary data to merge.}

\item{region_name}{A string specifying the column name in \code{region_data}
that contains the region names.}

\item{auxiliary_region_name}{A string specifying the column name in
\code{auxiliary_data} that contains the region names}

\item{csv_path}{A string specifying the path of the csv which has
alternative names for regins in its rows}

\item{distance}{An integer representing the maximum edit distance permitted}

\item{method}{A string specifing the edit distance algorithm to be used}

\item{basic}{A boolean representing whether the edit distance algorithm
should be used}
}
\value{
An \code{sf} data frame containing the merged data.
}
\description{
Merges two datasets based on values in a csv.
}
\examples{
merged_data <- merge_data_with_csv(country_data, auxiliary_data,
                                   "country", "country_name", csv_path)
}
