% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data_with_ui}
\alias{merge_data_with_ui}
\title{Merges two datasets based on region names, using name matching,
and prompts user input if the regions cannot be allocated automatically.}
\usage{
merge_data_with_ui(
  region_data,
  auxiliary_data,
  region_name,
  auxiliary_region_name,
  csv_path,
  distance = 3,
  method = "qgram"
)
}
\arguments{
\item{region_data}{A data frame containing the region data.}

\item{auxiliary_data}{A data frame containing the auxiliary data to merge.}

\item{region_name}{A string specifying the column name in \code{region_data}
that contains the region names.}

\item{auxiliary_region_name}{A string specifying the column name in
\code{auxiliary_data} that contains the region names}

\item{csv_path}{A string specifying the path of the csv which has
alternative names for regins in its rows}

\item{distance}{An integer representing the maximum edit distance permitted}

\item{method}{A string specifing the edit distance algorithm to be used}
}
\value{
An \code{sf} data frame containing the merged data.
}
\description{
Merges two datasets based on region names, using name matching,
and prompts user input if the regions cannot be allocated automatically.
}
\examples{
merged_data <- merge_data_with_ui(country_data, auxiliary_data,
                                  "country", "country_name", csv_path)
}
