% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{run_country_matching_ui}
\alias{run_country_matching_ui}
\title{Iterates through each country and checks whether it has been allocated
a number automatically. If not, it prompts the user to input a name or
whether the country should be removed.}
\usage{
run_country_matching_ui(country_data, country_name)
}
\arguments{
\item{country_data}{A data frame containing the country data.}

\item{country_name}{A string specifying the column name in \code{country_data}
that contains the country names.}
}
\value{
An \code{sf} data frame containing new country numbers
}
\description{
Iterates through each country and checks whether it has been allocated
a number automatically. If not, it prompts the user to input a name or
whether the country should be removed.
}
\examples{
country_data <- run_country_matching_ui(country_data, "country")
}
