% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shapefile_istat_regioni}
\alias{shapefile_istat_regioni}
\title{Shape Data of Italian Regions by ISTAT}
\format{Un data frame con 77950 osservazioni e 8 variabili. Ciascuna osservazione identifica un punto per costruire i confini tra le diverse regioni.}
\source{
\url{http://www.istat.it/it/archivio/24613}
}
\description{
Geometrie per disegnare le regioni italiane fornite da ISTAT.
}
\details{
Dati aggiornati al 1 gennaio 2011.

  Rilasciati sotto licenza Creative Commons - Attribuzione - versione 3.0. Per maggiori dettagli sulla licenza: http://creativecommons.org/licenses/by/3.0/it/
}

