% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialize_country_dataset.R
\name{spatialize_country_dataset}
\alias{spatialize_country_dataset}
\title{spatialize_country_dataset}
\usage{
spatialize_country_dataset(stats, by, variable, maptype, m49_codes_to_hide)
}
\value{
an object from \pkg{sf}
}
\description{
Spatializes a statistical country dataset
}
