% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialize_dataset.R
\name{spatialize_dataset}
\alias{spatialize_dataset}
\title{spatialize_dataset}
\usage{
spatialize_dataset(sf, sfby, sfby.code, stats, by, variable, maptype, m49_codes_to_hide)
}
\value{
an object from \pkg{sf}
}
\description{
Spatializes a statistical dataset
}
