% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_country_map.R
\name{create_country_map}
\alias{create_country_map}
\title{create_country_map}
\usage{
create_country_map(
  stats,
  by,
  variable,
  digits = 2,
  lang = "en",
  maptype = "choropleth",
  classtype = "jenks",
  classnumber = 5,
  breaks,
  col = "#08519C",
  pal = NULL,
  invertpal = FALSE,
  bgCol = "transparent",
  bgBorderCol = "transparent",
  naCol = "gray",
  naColBox = "lightgray",
  boundCol = "white",
  contCol = "lightgray",
  hashCol = "lightgray",
  m49_codes_to_hide = "010",
  add_small_features_as_dots = TRUE,
  add_small_NA_features_as_dots = FALSE,
  small_features_dots_cex = 0.4,
  legendtitle = "Legend",
  legendunit = "",
  legendcol = "black",
  add_disclaimers = TRUE,
  add_copyright = TRUE,
  add = FALSE,
  family = ""
)
}
\description{
Creates a country map
}
