% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{create_map}
\usage{
create_map(
  sf = NULL,
  sfby = NULL,
  sfby.code = NULL,
  stats = NULL,
  by = NULL,
  variable,
  digits = 2,
  lang = "en",
  maptype = "choropleth",
  classtype = "jenks",
  classnumber = 5,
  breaks,
  col = "#08519C",
  pal = NULL,
  invertpal = FALSE,
  bgCol = "transparent",
  bgBorderCol = "transparent",
  naCol = "gray",
  naColBox = "lightgray",
  boundCol = "white",
  contCol = "lightgray",
  hashCol = "lightgray",
  m49_codes_to_hide = "010",
  add_small_features_as_dots = TRUE,
  add_small_NA_features_as_dots = FALSE,
  small_features_dots_cex = 0.4,
  pch = 21,
  level.min = NULL,
  level.max = NULL,
  level.unit = "chars",
  plot.handler = NULL,
  legend = TRUE,
  legendtitle = "Legend",
  legendunit = "",
  legendcol = "black",
  legendpch = pch,
  legendpchcol = col,
  legend_nesting = FALSE,
  add_disclaimers = TRUE,
  add_copyright = TRUE,
  add = FALSE,
  family = "FuturaStd"
)
}
\description{
Creates a map
}
