% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lines.R
\name{find_closest_rects}
\alias{find_closest_rects}
\title{Finds the closest rectangle in free space for a small countries
information to be sotred in.}
\usage{
find_closest_rects(data, small_country_area, width, height)
}
\arguments{
\item{data}{The map data which should include simple features information.}

\item{small_country_area}{A value representing the threshold area below which
countries are considered small.}

\item{width}{The width of the free space the small countries
should be assigned}

\item{height}{The height of the free space the small countries
should be assigned}
}
\value{
A data frame containing the closest rectangle for each small country.
}
\description{
Finds the closest rectangle in free space for a small countries
information to be sotred in.
}
\examples{
# Assuming `world_data` is a valid `sf` object containing world map data:
closest_rects <- find_closest_rects(world_data,
 small_country_area = 100000, width = 1, height = 1)

}
