% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching.R
\name{get_region_number_with_edit_distance}
\alias{get_region_number_with_edit_distance}
\title{Returns the country number based off the levenstein distance being
less than x}
\usage{
get_region_number_with_edit_distance(
  country,
  distance,
  csv_path,
  method = "qgram"
)
}
\arguments{
\item{method}{A string representing the edit distance method}

\item{region}{A string representing the region name to search for.}

\item{x}{A numeric threshold for the Levenshtein distance.}
}
\value{
The row index of the closest matching country
if found; otherwise, -1.
}
\description{
This function searches for the country in the CSV data by calculating the
Levenshtein distance between the input \code{country} and the country names in
the data. If the distance is less than a threshold \code{x}, it returns the
index of the matching row which is the country number.
}
\examples{
country_number <- get_country_number_with_edit_distance("germany", 3)

}
