% Generated by roxygen2 (4.0.2): do not edit by hand
\name{map}
\alias{map}
\title{Easier choropleth maps}
\usage{
map(region = "Italy", ...)
}
\arguments{
\item{region}{The region for which the map should be drawn (currently only 'Italy')}

\item{...}{Arguments for the calle dfunction}
}
\description{
Easier choropleth maps
}
\author{
Nicola Sturaro
}

