% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{mapIT-package}
\alias{mapIT-package}
\title{Choropleth Maps of Italy}
\description{
Creazione di mappe (coroplete) delle regioni italiane in maniera semplice.
}
\details{
Al momento \code{\link{mapIT}} è l'unica funzione del package. \code{\link{map}} è un wrapper di \code{mapIT} per sviluppi futuri.
}
\examples{
# esempio ultra-minimale. Per la guida ed esempi si veda mapIT.
  mapIT(rnorm(20))
}
\author{
Nicola Sturaro Sommacal
}

