% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapit.R
\docType{package}
\name{mapit}
\alias{mapit}
\alias{mapit-package}
\title{Tools to produce UN standard maps}
\description{
Provides tools to automate production of UN standard maps while applying
 thematic cartography best practices.
}
\details{
\tabular{ll}{
   Package: \tab mapit\cr
   Type: \tab Package\cr
   Version: \tab 0.6\cr
   Date: \tab 2022-11-4\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eblondel/mapit}
  \item \url{https://eblondel.github.io/mapit/}
  \item Report bugs at \url{https://github.com/eblondel/mapit/issues}
}

}
\author{
\strong{Maintainer}: Emmanuel Blondel \email{emmanuel.blondel1@gmail.com} (\href{https://orcid.org/0000-0002-5870-5762}{ORCID})

Other contributors:
\itemize{
  \item Alexandre Bennici (\href{https://orcid.org/0000-0003-2160-3487}{ORCID}) [contributor]
}

}
