% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data_with_edit_distance}
\alias{merge_data_with_edit_distance}
\title{Merges two datasets based on region names using levenshtein distance.}
\usage{
merge_data_with_edit_distance(
  region_data,
  auxiliary_data,
  region_name,
  auxiliary_region_name,
  distance = 3,
  method = "lv"
)
}
\arguments{
\item{region_data}{A data frame containing the region data.}

\item{auxiliary_data}{A data frame containing the auxiliary data to merge.}

\item{region_name}{A string specifying the column name in \code{region_data}
that contains the region names.}

\item{auxiliary_region_name}{A string specifying the column name in
\code{auxiliary_data} that contains the region names}

\item{distance}{An integer representing the maximum levenshtein distance
between two strings which can be matched}
}
\value{
An \code{sf} data frame containing the merged data.
}
\description{
Merges two datasets based on region names using levenshtein distance.
}
\examples{
merged_data <- merge_data_with_levenshtein_distance(country_data,
                                                    auxiliary_data,
                                                    "country",
                                                    "country_name", 3)
}
