% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data_with_ui}
\alias{merge_data_with_ui}
\title{Merges two datasets based on country names, using name matching,
and prompts user input if the countries cannot be allocated automatically.}
\usage{
merge_data_with_ui(
  country_data,
  auxiliary_data,
  country_name,
  auxiliary_country_name
)
}
\arguments{
\item{country_data}{A data frame containing the country data.}

\item{auxiliary_data}{A data frame containing the auxiliary data to merge.}

\item{country_name}{A string specifying the column name in \code{country_data}
that contains the country names.}

\item{auxiliary_country_name}{A string specifying the column name in
\code{auxiliary_data} that contains the country names}
}
\value{
An \code{sf} data frame containing the merged data.
}
\description{
Merges two datasets based on country names, using name matching,
and prompts user input if the countries cannot be allocated automatically.
}
\examples{
merged_data <- merge_data_with_ui(country_data, auxiliary_data,
                                  "country", "country_name")
}
