% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lines.R
\name{modify_label_positions}
\alias{modify_label_positions}
\title{Modifies Label Positions Based on the closest rectangles if the countries
area is too small}
\usage{
modify_label_positions(
  data,
  small_country_area,
  width,
  height,
  label_x = "label_x",
  label_y = "label_y"
)
}
\arguments{
\item{data}{The map data which should include simple features information.}

\item{small_country_area}{A value representing the threshold area below which
countries are considered small.}

\item{width}{The width of the free space the small countries
should be assigned}

\item{height}{The height of the free space the small countries
should be assigned}

\item{label_x}{The column name for the x-coordinate of country labels.}

\item{label_y}{The column name for the y-coordinate of country labels.}
}
\value{
A data frame containing the original
data along with updated label positions.
}
\description{
Modifies Label Positions Based on the closest rectangles if the countries
area is too small
}
\examples{
# Assuming `world_data` is a valid `sf` object containing world map data:
modified_labels <- modify_label_positions(world_data,
                                          small_country_area = 100000,
                                          width = 1, height = 1)

}
