\name{shapefile_istat_regioni}
\alias{shapefile_istat_regioni}
\alias{shapedata_istat_regioni}
\docType{data}

\title{Shape Data of Italian Regions by ISTAT}

\description{
Geometrie per disegnare le regioni italiane fornite da ISTAT.
}

\usage{data(shapefile_istat_regioni)}

\format{
  Un data frame con 77950 osservazioni e 8 variabili. Ciascuna osservazione identifica un punto per costruire i confini tra le diverse regioni.
}

\details{
Dati aggiornati al 1 gennaio 2011.

Rilasciati sotto licenza Creative Commons - Attribuzione - versione 3.0. Per maggiori dettagli sulla licenza: http://creativecommons.org/licenses/by/3.0/it/
}

\source{
http://www.istat.it/it/archivio/24613
}

\examples{
data(shapefile_istat_regioni)
}

\keyword{datasets}
